+++ draft-packages/stories/TitleBlock.stories.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@kaizen/component-library/components/Button"
+import { Button } from "@kaizen/draft-button"
 import { TitleBlock } from "@kaizen/draft-title-block"
 import * as React from "react"