+++ draft-packages/zen-off-canvas/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.6.12...@kaizen/draft-zen-off-canvas@1.6.13) (2020-06-10)
+
+**Note:** Version bump only for package @kaizen/draft-zen-off-canvas
+
+
+
+
+
+## [1.6.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.6.11...@kaizen/draft-zen-off-canvas@1.6.12) (2020-06-09)
+
+**Note:** Version bump only for package @kaizen/draft-zen-off-canvas
+
+
+
+
+
 ## [1.6.11](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.6.10...@kaizen/draft-zen-off-canvas@1.6.11) (2020-06-05)
 
 **Note:** Version bump only for package @kaizen/draft-zen-off-canvas