+++ packages/component-library/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [7.21.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@7.20.0...@kaizen/component-library@7.21.0) (2020-06-10)
+
+
+### Features
+
+* Add blur and focus callbacks to form components ([#537](https://github.com/cultureamp/kaizen-design-system/issues/537)) ([5bd717b](https://github.com/cultureamp/kaizen-design-system/commit/5bd717b019d89da27cee0cd4b198c606b458b795))
+
+
+
+
+
+# [7.20.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@7.19.3...@kaizen/component-library@7.20.0) (2020-06-09)
+
+
+### Features
+
+* Rollout inter ([#530](https://github.com/cultureamp/kaizen-design-system/issues/530)) ([a1b2059](https://github.com/cultureamp/kaizen-design-system/commit/a1b2059980ea753036caa5cb15ba6b1235d52ba4))
+
+
+
+
+
 ## [7.19.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/component-library@7.19.2...@kaizen/component-library@7.19.3) (2020-06-05)
 
 **Note:** Version bump only for package @kaizen/component-library